Get help view controller from shared method
